Remove numerical tolerance checks in neff.py in ne function
Experimental change: not sure what these checks were for,
but they affected behavior of optimizer for soft_optimized,
which gave wrong results.

Removing the checks and resetting weights helped the optimizer.
No malfunctions of the code was found.
but I am not sure if it breaks other code.
If there is a problem, more careful analysis should be done.